refactor: rename permissions mod to access_control
