+++ src/sequence/sequence.rs
@@ -7,7 +7,7 @@
 // specific language governing permissions and limitations relating to use of the SAFE Network
 // Software.
 
-use crate::permissions::{
+use crate::access_control::{
     DataPermissions, PrivatePermissionSet, PrivatePermissions, PublicPermissionSet,
     PublicPermissions, Request,
 };
